feat:cost control

+++ api/v1/costControl/service.go
@@ -5,6 +5,8 @@ import (
 	"bpm/api/v1/position"
 	"bpm/api/v1/project"
 	"bpm/core/database"
+	"bpm/core/queue"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"strconv"
@@ -214,18 +216,55 @@ func (s *costControlService) NewPaymentRequest(info ReqPaymentRequestNew) error 
 			return errors.New(msg)
 		}
 	}
+	for _, audit := range *auditInfo {
+		var auditNew ReqPaymentRequestAuditNew
+		auditNew.PaymentRequestID = id
+		auditNew.AuditLevel = audit.AuditLevel
+		auditNew.AuditType = audit.AuditType
+		auditNew.AuditTo = audit.AuditTo
+		auditNew.User = info.User
+		err = repo.CreatePaymentRequestAudit(auditNew)
+		if err != nil {
+			msg := "创建请款记录审核失败"
+			return errors.New(msg)
+		}
+	}
 	var history ReqPaymentRequestHistoryNew
 	history.PaymentRequestID = id
 	history.OrganizationID = info.OrganizationID
 	history.User = info.User
 	history.Action = "新建"
+	history.Content = ""
 	history.Remark = "请款已新建，当前状态为待审核"
-	err = repo.CreatePaymentRequestHistory(history)
+	historyID, err := repo.CreatePaymentRequestHistory(history)
 	if err != nil {
 		msg := "创建请款历史失败"
 		return errors.New(msg)
 	}
+	for _, picture := range info.Picture {
+		var pictureInfo ReqPaymentRequestHistoryPictureNew
+		pictureInfo.PaymentRequestHistoryID = historyID
+		pictureInfo.Picture = picture
+		pictureInfo.User = info.User
+		err = repo.CreatePaymentRequestHistoryPicture(pictureInfo)
+		if err != nil {
+			msg := "创建请款记录历史图片失败"
+			return errors.New(msg)
+		}
+	}
 	tx.Commit()
+	type NewPaymentRequestCreated struct {
+		PaymentRequestID int64 `json:"payment_request_id"`
+	}
+	var newEvent NewPaymentRequestCreated
+	newEvent.PaymentRequestID = id
+	rabbit, _ := queue.GetConn()
+	msg, _ := json.Marshal(newEvent)
+	err = rabbit.Publish("NewPaymentRequestCreated", msg)
+	if err != nil {
+		msg := "create event NewPaymentRequestCreated error"
+		return errors.New(msg)
+	}
 	return nil
 
 }
@@ -311,10 +350,22 @@ func (s *costControlService) UpdatePaymentRequest(info ReqPaymentRequestUpdate, 
 	history.OrganizationID = oldPaymentRequest.OrganizationID
 	history.User = info.User
 	history.Action = "更新"
+	history.Content = ""
 	history.Remark = "请款已更新，当前状态为待审核"
-	err = repo.CreatePaymentRequestHistory(history)
+	historyID, err := repo.CreatePaymentRequestHistory(history)
 	if err != nil {
 		return err
+	}
+	for _, picture := range info.Picture {
+		var pictureInfo ReqPaymentRequestHistoryPictureNew
+		pictureInfo.PaymentRequestHistoryID = historyID
+		pictureInfo.Picture = picture
+		pictureInfo.User = info.User
+		err = repo.CreatePaymentRequestHistoryPicture(pictureInfo)
+		if err != nil {
+			msg := "创建请款记录历史图片失败"
+			return errors.New(msg)
+		}
 	}
 	tx.Commit()
 	return nil
@@ -337,6 +388,11 @@ func (s *costControlService) GetPaymentRequestList(filter ReqPaymentRequestFilte
 			return 0, nil, err
 		}
 		(*list)[key].Picture = *pictures
+		audits, err := query.GetPaymentRequestAuditList(budget.ID)
+		if err != nil {
+			return 0, nil, err
+		}
+		(*list)[key].Audit = *audits
 	}
 	return count, list, nil
 }
@@ -344,11 +400,11 @@ func (s *costControlService) GetPaymentRequestList(filter ReqPaymentRequestFilte
 func (s *costControlService) GetPaymentRequestByID(id, organizationID int64) (*RespPaymentRequest, error) {
 	db := database.InitMySQL()
 	query := NewCostControlQuery(db)
-	budget, err := query.GetPaymentRequestByID(id)
+	paymentRequest, err := query.GetPaymentRequestByID(id)
 	if err != nil {
 		return nil, err
 	}
-	if budget.OrganizationID != organizationID && organizationID != 0 {
+	if paymentRequest.OrganizationID != organizationID && organizationID != 0 {
 		msg := "预算记录不存在或无权限"
 		return nil, errors.New(msg)
 	}
@@ -356,8 +412,13 @@ func (s *costControlService) GetPaymentRequestByID(id, organizationID int64) (*R
 	if err != nil {
 		return nil, err
 	}
-	budget.Picture = *pictures
-	return budget, nil
+	audits, err := query.GetPaymentRequestAuditList(id)
+	if err != nil {
+		return nil, err
+	}
+	paymentRequest.Picture = *pictures
+	paymentRequest.Audit = *audits
+	return paymentRequest, nil
 }
 
 func (s *costControlService) DeletePaymentRequest(id, organizationID int64, user string, userID int64) error {
@@ -381,6 +442,10 @@ func (s *costControlService) DeletePaymentRequest(id, organizationID int64, user
 		msg := "只能删除自己创建的请款"
 		return errors.New(msg)
 	}
+	if oldPaymentRequest.Status != 1 && oldPaymentRequest.Status != 2 && oldPaymentRequest.Status != 3 {
+		msg := "请款记录无法删除，可能已付款"
+		return errors.New(msg)
+	}
 	err = repo.DeletePaymentRequest(id, user)
 	if err != nil {
 		msg := "删除请款失败"
@@ -390,6 +455,23 @@ func (s *costControlService) DeletePaymentRequest(id, organizationID int64, user
 	if err != nil {
 		msg := "删除请款图片失败"
 		return errors.New(msg)
+	}
+
+	if oldPaymentRequest.BudgetID != 0 {
+		oldBudget, err := repo.GetBudgetByID(oldPaymentRequest.BudgetID)
+		if err != nil {
+			msg := "获取预算失败"
+			return errors.New(msg)
+		}
+		var budgetUpdate ReqBudgetPaid
+		budgetUpdate.Used = oldBudget.Used - oldPaymentRequest.Total
+		budgetUpdate.Balance = oldBudget.Balance + oldPaymentRequest.Total
+		budgetUpdate.User = user
+		err = repo.UpdateBudgitUsed(oldPaymentRequest.BudgetID, budgetUpdate)
+		if err != nil {
+			msg := "更新预算信息失败"
+			return errors.New(msg)
+		}
 	}
 	tx.Commit()
 	return nil
@@ -472,4 +554,289 @@ func (s *costControlService) GetPaymentRequestTypeList(filter ReqPaymentRequestT
 	type2.Audit = *res2
 	res := &[]RespPaymentRequestType{type1, type2}
 	return res, err
+}
+
+func (s *costControlService) AuditPaymentRequest(paymentRequestID int64, info ReqPaymentRequestAudit) error {
+	db := database.InitMySQL()
+	tx, err := db.Begin()
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+	repo := NewCostControlRepository(tx)
+	paymentRequest, err := repo.GetPaymentRequestByID(paymentRequestID)
+	if err != nil {
+		msg := "获取请款记录失败"
+		return errors.New(msg)
+	}
+	if paymentRequest.Status != 1 {
+		msg := "此请款无法审核"
+		return errors.New(msg)
+	}
+	assignExist, err := repo.CheckAudit(paymentRequestID, info.UserID, info.PositionID, paymentRequest.AuditLevel)
+	fmt.Println(info.PositionID, paymentRequest.AuditLevel)
+	if err != nil {
+		msg := "检查审核设置失败"
+		return errors.New(msg)
+	}
+	if assignExist == 0 {
+		msg := "此请款审核未分配给你"
+		return errors.New(msg)
+	}
+	nextLevel := 0
+	result := "审核通过"
+	remark := "审核已通过"
+	status := 2
+	if info.Result == 1 {
+		nextLevel, err = repo.GetNextLevel(paymentRequestID, paymentRequest.AuditLevel)
+		if err != nil {
+			if err.Error() == "sql: no rows in result set" {
+				nextLevel = 0
+				remark += "，没有下一层审核，当前状态为审核通过"
+			} else {
+				msg := "获取下一层审核失败"
+				return errors.New(msg)
+			}
+		} else {
+			remark += "，当前状态为待审核，下一层审核为第" + fmt.Sprintf("%d", nextLevel) + "层"
+			status = 1
+		}
+	} else {
+		result = "审核不通过"
+		remark = "审核不通过，当前状态为审核驳回"
+		status = 3
+		nextLevel = 1
+	}
+	// remark += "\n " + info.Content
+	var history ReqPaymentRequestHistoryNew
+	history.PaymentRequestID = paymentRequestID
+	history.OrganizationID = paymentRequest.OrganizationID
+	history.User = info.User
+	history.Action = result
+	history.Content = info.Content
+	history.Remark = remark
+	historyID, err := repo.CreatePaymentRequestHistory(history)
+	if err != nil {
+		return err
+	}
+	err = repo.AuditPaymentRequest(paymentRequestID, nextLevel, status, info.User)
+	if err != nil {
+		msg := "更新请款状态失败"
+		return errors.New(msg)
+	}
+	for _, link := range info.File {
+		var paymentRequestHistoryPicture ReqPaymentRequestHistoryPictureNew
+		paymentRequestHistoryPicture.PaymentRequestHistoryID = historyID
+		paymentRequestHistoryPicture.Picture = link
+		paymentRequestHistoryPicture.User = info.User
+		err = repo.CreatePaymentRequestHistoryPicture(paymentRequestHistoryPicture)
+		if err != nil {
+			msg := "创建文件失败"
+			return errors.New(msg)
+		}
+	}
+	if nextLevel == 0 && paymentRequest.BudgetID != 0 {
+		oldBudget, err := repo.GetBudgetByID(paymentRequest.BudgetID)
+		if err != nil {
+			msg := "获取预算失败"
+			return errors.New(msg)
+		}
+		var budgetUpdate ReqBudgetPaid
+		budgetUpdate.Used = oldBudget.Used + paymentRequest.Total
+		budgetUpdate.Balance = oldBudget.Balance - paymentRequest.Total
+		budgetUpdate.User = info.User
+		err = repo.UpdateBudgitUsed(paymentRequest.BudgetID, budgetUpdate)
+		if err != nil {
+			msg := "更新预算信息失败"
+			return errors.New(msg)
+		}
+	}
+
+	tx.Commit()
+	type NewPaymentRequestAudited struct {
+		PaymentRequestID int64 `json:"paymentRequest_id"`
+	}
+	var newPaymentRequest NewPaymentRequestAudited
+	newPaymentRequest.PaymentRequestID = paymentRequestID
+	rabbit, _ := queue.GetConn()
+	msg, _ := json.Marshal(newPaymentRequest)
+	err = rabbit.Publish("NewPaymentRequestAudited", msg)
+	if err != nil {
+		msg := "create paymentRequest NewPaymentRequestAudited error"
+		return errors.New(msg)
+	}
+	return nil
+}
+
+func (s *costControlService) GetPaymentRequestHistoryList(filter ReqPaymentRequestHistoryFilter) (*[]RespPaymentRequestHistory, error) {
+
+	db := database.InitMySQL()
+	query := NewCostControlQuery(db)
+	paymentRequest, err := query.GetPaymentRequestByID(filter.PaymentRequestID)
+	if err != nil {
+		msg := "获取请款记录失败"
+		return nil, errors.New(msg)
+	}
+	if paymentRequest.OrganizationID != filter.OrganizationID && filter.OrganizationID != 0 {
+		msg := "请款记录不存在"
+		return nil, errors.New(msg)
+	}
+	list, err := query.GetPaymentRequestHistoryList(filter.PaymentRequestID)
+	if err != nil {
+		return nil, err
+	}
+	for key, history := range *list {
+		pictures, err := query.GetPaymentRequestHistoryPictureList(history.ID)
+		if err != nil {
+			return nil, err
+		}
+		(*list)[key].Picture = *pictures
+	}
+	return list, nil
+}
+
+func (s *costControlService) UpdatePaymentRequestAudit(id int64, info ReqPaymentRequestAuditUpdate, organizationID int64) error {
+	db := database.InitMySQL()
+	tx, err := db.Begin()
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+	repo := NewCostControlRepository(tx)
+	positionRepo := position.NewPositionRepository(tx)
+	userRepo := auth.NewAuthRepository(tx)
+	err = repo.DeletePaymentRequestAudit(id, organizationID, info.User)
+	if err != nil {
+		msg := "更新请款审核失败"
+		return errors.New(msg)
+	}
+	for _, audit := range info.AuditInfo {
+		for _, auditTo := range audit.AuditTo {
+			var auditInfo ReqPaymentRequestAuditNew
+			auditInfo.PaymentRequestID = id
+			auditInfo.AuditLevel = audit.AuditLevel
+			auditInfo.AuditType = audit.AuditType
+			auditInfo.AuditTo = auditTo
+			if auditInfo.AuditType == 1 {
+				_, err := positionRepo.GetPositionByID(auditTo, organizationID)
+				if err != nil {
+					msg := "审核层次" + strconv.Itoa(int(auditInfo.AuditLevel)) + "职位不存在"
+					return errors.New(msg)
+				}
+			} else {
+				userInfo, err := userRepo.GetUserByID(auditTo)
+				if err != nil {
+					msg := "审核层次" + strconv.Itoa(int(auditInfo.AuditLevel)) + "用户不存在"
+					return errors.New(msg)
+				}
+				if userInfo.OrganizationID != organizationID {
+					msg := "审核层次" + strconv.Itoa(int(auditInfo.AuditLevel)) + "用户不存在"
+					return errors.New(msg)
+				}
+			}
+			auditInfo.User = info.User
+			err = repo.CreatePaymentRequestAudit(auditInfo)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	err = repo.AuditPaymentRequest(id, 1, 1, info.User)
+	if err != nil {
+		msg := "更新请款审核失败"
+		return errors.New(msg)
+	}
+	tx.Commit()
+	return nil
+
+}
+
+func (s *costControlService) NewPayment(paymentRequestID int64, info ReqPaymentNew, organizationID int64) error {
+	db := database.InitMySQL()
+	tx, err := db.Begin()
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+	repo := NewCostControlRepository(tx)
+	paymentRequest, err := repo.GetPaymentRequestByID(paymentRequestID)
+	if err != nil {
+		msg := "获取请款记录失败"
+		return errors.New(msg)
+	}
+	if paymentRequest.OrganizationID != organizationID {
+		msg := "请款记录不存在"
+		return errors.New(msg)
+	}
+	if paymentRequest.Status != 2 && paymentRequest.Status != 4 {
+		msg := "请款记录状态不正确"
+		return errors.New(msg)
+	}
+	if info.Amount > paymentRequest.Due {
+		msg := "此次付款金额大于未付款金额"
+		return errors.New(msg)
+	}
+	info.OrganizationID = paymentRequest.OrganizationID
+	info.ProjectID = paymentRequest.ProjectID
+	paymentID, err := repo.CreatePayment(paymentRequestID, info)
+	if err != nil {
+		msg := "生成付款记录失败"
+		return errors.New(msg)
+	}
+	for _, picture := range info.Picture {
+		var paymentPicture ReqPaymentPictureNew
+		paymentPicture.PaymentID = paymentID
+		paymentPicture.Picture = picture
+		paymentPicture.User = info.User
+		err = repo.CreatePaymentPicture(paymentPicture)
+		if err != nil {
+			msg := "创建付款记录文件失败"
+			return errors.New(msg)
+		}
+	}
+	var paymentRequestUpdate ReqPaymentRequestPaid
+	paymentRequestUpdate.Paid = paymentRequest.Paid + info.Amount
+	paymentRequestUpdate.Due = paymentRequest.Due - info.Amount
+	if paymentRequestUpdate.Due == 0 {
+		paymentRequestUpdate.Status = 5
+	} else {
+		paymentRequestUpdate.Status = 4
+	}
+	paymentRequestUpdate.User = info.User
+	err = repo.UpdatePaymentRequestPaid(paymentRequestID, paymentRequestUpdate)
+	if err != nil {
+		msg := "更新请款信息失败"
+		return errors.New(msg)
+	}
+	var history ReqPaymentRequestHistoryNew
+	history.PaymentRequestID = paymentRequestID
+	history.OrganizationID = paymentRequest.OrganizationID
+	history.User = info.User
+	history.Action = "付款"
+	history.Remark = "本次付款金额" + strconv.FormatFloat(info.Amount, 'f', 2, 64) + "元，"
+	if paymentRequestUpdate.Due == 0 {
+		history.Remark += "已完全付款，当前状态为已付款"
+	} else {
+		history.Remark += "未完全付款，当前状态为部分付款"
+	}
+	history.Content = info.Remark
+	historyID, err := repo.CreatePaymentRequestHistory(history)
+	if err != nil {
+		msg := "生成付款记录失败"
+		return errors.New(msg)
+	}
+	for _, link := range info.Picture {
+		var paymentRequestHistoryPicture ReqPaymentRequestHistoryPictureNew
+		paymentRequestHistoryPicture.PaymentRequestHistoryID = historyID
+		paymentRequestHistoryPicture.Picture = link
+		paymentRequestHistoryPicture.User = info.User
+		err = repo.CreatePaymentRequestHistoryPicture(paymentRequestHistoryPicture)
+		if err != nil {
+			msg := "创建付款记录文件失败"
+			return errors.New(msg)
+		}
+	}
+	tx.Commit()
+	return nil
+
 }
